refactor search info by name
