+++ packages/lib/src/lib/search/index.ts
@@ -1,2 +1,3 @@
 export { initAddresses, searchAddress } from './address'
+export { makeGetInfoByName } from './info'
 export { type SearchContext, type SearchPos } from './types'
